aosqa_ders_records-main/requirements.txt
googletrans==3.1.0a0
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
html5lib==1.0.1
httpcore==0.9.1
httplib2==0.14.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.8
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonfield==3.1.0
jsonschema==4.1.0
Pillow==7.0.0
platformdirs==2.4.0
playwright==1.18.2
prettytable==3.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2-binary==2.8.6

pyrsistent==0.18.0
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyxattr==0.6.1
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
reportlab==3.5.32
requests==2.27.1
requests-unixsocket==0.2.0
rfc3986==1.5.0


systemd-python==234
tablib==3.0.0
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.0
trio==0.19.0
wadllib==1.3.3
wcwidth==0.2.5
webdriver-manager==3.5.3
webencodings==0.5.1
websockets==10.1
Werkzeug==2.0.3
widgetsnbextension==3.5.1
wsproto==1.0.0
xhtml2pdf==0.2.3
xlrd==2.0.1
xlwt==1.3.0

